+++ app/app/group/[id]/components/buttons/ApplicationRegisterNewGroupButton.jsx
