+++ app/app/group/[id]/components/modals/StudentDeleteModalDialog.jsx
